Added NotifyIcon
got source from here:
http://www.codeproject.com/Articles/36468/WPF-NotifyIcon
the code is delivered under the CPOL license
http://www.codeproject.com/info/cpol10.aspx
which is the reason why I didn't change the namespace

CPOL extractions:
3. d. You may distribute the standard version of the Executable Files
and Source Code or Derivative Work in aggregate with other (possibly
commercial) programs as part of a larger (possibly commercial) software
distribution.

5.d. You agree not to sell, lease, or rent any part of the Work. This
does not restrict you from including the Work or any part of the Work
inside a larger software distribution that itself is being sold. The
Work by itself, though, cannot be sold, leased or rented.

5.f. You agree not to use the Work for illegal, immoral or improper
purposes, or on pages containing illegal, immoral or improper material.
The Work is subject to applicable export laws. You agree to comply with
all such laws and regulations that may apply to the Work after Your
receipt of the Work.